refactored to run selected actions
